requirements.txt
numpy==1.24.3
PyYAML==6.0
streamlit==1.21.0
torch==2.0.0
